% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnEnC2R.R
\name{AnEnC2R}
\alias{AnEnC2R}
\title{RAnEn::AnEnC2R}
\usage{
AnEnC2R(AnEn, member)
}
\arguments{
\item{AnEn}{An AnEn class object.}

\item{member}{The name of member to convert indices.}
}
\value{
An AnEn class object.
}
\description{
RAnEn::AnEnC2R converts the C indices to R indices. This function is not
for the general use.
}
\author{
Weiming Hu \email{weiming@psu.edu}
}
