% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasCorrection.R
\name{biasCorrection}
\alias{biasCorrection}
\title{RAnEn::biasCorrection}
\usage{
biasCorrection(analogs = stop("Please set analogs"),
  observations = stop("Please set observations"),
  forecasts = stop("Please set forecasts"),
  forecast.ID = stop("Please set forecast.ID"),
  test.ID.start = stop("Please set test.ID.start"), group.func = mean,
  ..., keep.bias = F, show.progress = T)
}
\arguments{
\item{analogs}{analogs should either be the class of 'AnEn' or a 
5-dimensional array. A variable of the class 'AnEn' is generated
from the function \code{\link{compute_analogs}}.}

\item{observations}{A 4-dimensional double array that contains 
all the observation parameter values. The expected structure is 
observations[parameter, station, day, flt] where flt stands for 
forecast lead time.}

\item{forecasts}{A 4-dimensional double array that contains all
the forecast parameter values. The expected structure is 
forecasts [parameter, station, day, flt] where flt stands for
forecast lead time.}

\item{forecast.ID}{A positive integer specifying which variable in
forecasts are used to compute the bias.}

\item{test.ID.start}{A positive integer specifying the start day
of tests.}

\item{group.func}{A function that summarizes all the member values
into one value. For example, it can be 'mean', 'median', 'max', or
'min'.}

\item{...}{Extra parameters to function group.func.}

\item{keep.bias}{A logical for whether to keep the bias array}

\item{show.progress}{A logical for whether to show the progress bar}
}
\value{
An array or a list depending on whether keep.bias is selected.
}
\description{
RAnEn::biasCorrection
}
\author{
Weiming Hu \email{weiming@psu.edu}

Statical bias correction function for analog ensembles.
}
