% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasCorrectionInsitu.R
\name{biasCorrectionInsitu}
\alias{biasCorrectionInsitu}
\title{RAnEn::biasCorrectionInsitu}
\usage{
biasCorrectionInsitu(
  AnEn,
  config,
  forecast.ID,
  group.func = mean,
  ...,
  keep.bias = F,
  show.progress = F,
  overwrite = F,
  density.limit = 1
)
}
\arguments{
\item{AnEn}{An AnEn object.}

\item{config}{The configuration object used to generate the AnEn results.}

\item{forecast.ID}{Which parameter in the forecasts should be used to generate bias.}

\item{group.func}{A function that summarizes all the member values
into one value. For example, it can be 'mean', 'median', 'max', or 'min'.}

\item{...}{Extra parameters to function group.func.}

\item{keep.bias}{A logical for whether to keep the bias array}

\item{show.progress}{A logical for whether to show the progress bar}

\item{overwrite}{A logical for whether to overwrite correction results.}

\item{density.limit}{The probability density threshold. If the density of the bin including
the current forecast is smaller than or equals to this value, the current forecast will be bias
corrected. This parameter is useful when only a portion of the test
forecasts need to be bias corrected.}
}
\value{
An AnEn object.
}
\description{
Statistical bias correction function for analog ensembles. This method uses the bias of the 
forecast model to correct AnEn, therefore the word 'in situ'.
}
\details{
For a given forecast from AnEn for a certain station/grid point, day, and FLT, first,
the mean of the selected forecasts, also the most similar past forecasts, is calculated;
then the difference of the mean of the selected forecasts and the current forecast is calculated.
This difference is then added to the AnEn forecasts.

This method only works when the vairable forecasted by AnEn also exists in the model forecasts.
For example, temperature is a model forecast variable. So we can bias correct AnEn forecasts for
temperature using the model forecasts. If the AnEn forecast variable does not exist in the model,
this method is not applicable.
}
\author{
Weiming Hu \email{weiming@psu.edu}
}
