% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasCorrectionNeuralNet.R
\name{biasCorrectionNeuralNet}
\alias{biasCorrectionNeuralNet}
\title{RAnEn::biasCorrectionNeuralNet}
\usage{
biasCorrectionNeuralNet(
  AnEn,
  config,
  sample.size = 20,
  sample.method = "systematic",
  cor.threshold = 0.1,
  maxit = 2000,
  nnet.size = NA,
  group.func,
  ...,
  show.progress = F,
  keep.bias = F,
  keep.model = F,
  overwrite = F,
  parallel = F,
  num.cores = stop("Please specify num.cores!")
)
}
\arguments{
\item{AnEn}{An AnEn object.}

\item{config}{The configuration object used to generate the AnEn results.}

\item{sample.size}{The number of samples to take from search days.}

\item{sample.method}{The sample method. It can be one of 'systematic' and 'random'.}

\item{cor.threshold}{The correlation threshold to determine whether a variable will
be used.}

\item{maxit}{The maximum number of iteration when training the neural network.}

\item{nnet.size}{The number of nodes in the hidden layer of the neural network.}

\item{group.func}{A function that summarizes all the member values
into one value. For example, it can be 'mean', 'median', 'max', or min'.}

\item{...}{Extra parameters to function group.func.}

\item{show.progress}{A logical for whether to show the progress bar}

\item{keep.bias}{A logical for whether to keep the bias array}

\item{keep.model}{A logical for whether to keep the neural network and data}

\item{overwrite}{A logical for whether to overwrite correction results.}

\item{parallel}{A logical for whether to use parallelization.}

\item{num.cores}{The number of cores to use for parallelization.}
}
\value{
An AnEn object.
}
\description{
Use a neural network to correct AnEn bias.
}
\details{
First, a group of search day samples are selected, either systematically or
randomly.

Then, analogs are generated for these sample days at all available locations
and FLTs using leave-one-out method. Leave-one-out method remoev only the sample
day from the search data set and use the rest of the search data as search.

After analogs are generated for samples, analog means are calculated and their
individual bias can be calculated as well.

Then, a neural network is trained potentially using the following variables:
\itemize{
  \item all variables from the model;
  \item locations if provided or loaction indices;
  \item FLT indices;
}

The output of the neural network will be bias. Which variables to use depends on
the correlation between variables and bias. the threshold can be set using
cor.threshold.

Finally, the bias for tests are calculated using the neural network and AnEn 
results are corrected.
}
\author{
Weiming Hu \email{weiming@psu.edu}
}
