% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.package.R
\name{check.package}
\alias{check.package}
\title{RAnEn::check.package}
\usage{
check.package(name)
}
\description{
RAnEn::check.package is the internal functions checking whether a
package is installed or not.
}
\keyword{internal}
