% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkOpenMP.R
\name{checkOpenMP}
\alias{checkOpenMP}
\title{RAnEn::checkOpenMP}
\usage{
checkOpenMP(silent = F)
}
\arguments{
\item{silent}{Whether to suppress output.}
}
\value{
A boolean whether OpenMP is supported.
}
\description{
This function checks whether OpenMP is supported in RAnEn library.
}
\author{
Weiming Hu \email{weiming@psu.edu}
}
