% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_analogs.R
\name{compute_analogs}
\alias{compute_analogs}
\title{RAnEn::compute_analogs}
\usage{
compute_analogs(forecasts, observations, circulars = vector(mode =
  "numeric"), weights = NA, stations_ID = NA, test_ID_start,
  test_ID_end, train_ID_start = 1, train_ID_end = NA,
  observation_ID = NA, members_size, rolling = 0, quick = T,
  cores = 0, search_extension = F, xs = NA, ys = NA,
  distance = 0, num_nearest = 0,
  observation_from_extended_station = F,
  recompute_sd_for_extended_station = F, output_search_stations = F,
  output_metric = F, verbose = 1, parameter_ID = NA,
  max_par_nan = 0, max_flt_nan = 0)
}
\description{
Deprecated function. Please see \code{\link{generateAnalogs}}.
}
