% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeWindDir.R
\name{computeWindDir}
\alias{computeWindDir}
\title{RAnEn::computeWindDir}
\usage{
computeWindDir(U, V)
}
\arguments{
\item{U}{A vector of values for U component of wind.}

\item{V}{A vector of values for V component of wind.}
}
\value{
A vector of values for wind speed.
}
\description{
Compute wind direction from U and V component of wind.
}
\author{
Weiming Hu \email{weiming@psu.edu}, Guido Cervone \email{cervone@psu.edu}
}
