% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeWindSpeed.R
\name{computeWindSpeed}
\alias{computeWindSpeed}
\title{RAnEn::computeWindSpeed}
\usage{
computeWindSpeed(U, V)
}
\arguments{
\item{U}{A vector of values for U component of wind.}

\item{V}{A vector of values for V component of wind.}
}
\value{
A vector of values for wind speed.
}
\description{
RAnEn::computeWindSpeed
}
\author{
Weiming Hu \email{weiming@psu.edu}
        Guido Cervone \email{cervone@psu.edu}

Compute wind speed from U and V component of wind.
}
