% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Config}
\alias{Config}
\title{Exposed C++ Class Config}
\description{
Analog generation can be fine-tuned with configuration. This
class provides all the tunable parameters for analog generation. You can
type \code{RAnEn::Config} in the R console to see the description for each parameter.
To create an object of Config, you can type \code{config <- new(Config)}, and then
you can directly modify the members, e.g. \code{config$num_analogs <- 10}. To see
the full list of tunable parameters, you can type \code{config} after you have
created an object. It will print all the parameters.
}
