% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToAdvance.R
\name{convertToAdvance}
\alias{convertToAdvance}
\title{RAnEn::convertToAdvance}
\usage{
convertToAdvance(config)
}
\arguments{
\item{configuration}{A configuration object generated by the function
\code{\link{generateConfiguration}}.}
}
\value{
An advanced configuration.
}
\description{
RAnEn::convertToAdvance converts a configuration to an advanced one.
}
\author{
Weiming Hu \email{weiming@psu.edu}
}
