% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatBasic.R
\name{formatBasic}
\alias{formatBasic}
\title{RAnEn::formatBasic}
\usage{
formatBasic(
  obj,
  df,
  col.par,
  circular.pars,
  col.x,
  col.y,
  col.station.name,
  time.series
)
}
\description{
RAnEn::formatBasic is an internal function called by RAnEn::formatObservations
and RAnEn::formatForecasts to convert a data frame to AnEn Observations or
Forecasts.
}
\keyword{internal}
