% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatConfiguration.R
\name{formatConfiguration}
\alias{formatConfiguration}
\title{RAnEn::formatConfiguration}
\usage{
formatConfiguration(config, verbose = T)
}
\arguments{
\item{config}{A configuration object generated by the function
\code{\link{generatedConfiguration}}.}

\item{verbose}{Whether to print progress information.}
}
\value{
The formatted configuration.
}
\description{
RAnEn::formatConfiguration formats configuration memebers.
}
\details{
Sometimes, a configuration member can be automatically formatted. For example,
for time variables, R has a POSIXct time but it is represented as numerics in
RAnEn. So this can be automicatially converted internally. This function
checks for those possibility to format configuration variables.
}
\author{
Weiming Hu \email{weiming@psu.edu}
}
