% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatForecasts.R
\name{formatForecasts}
\alias{formatForecasts}
\title{RAnEn::formatForecasts}
\usage{
formatForecasts(
  df,
  col.par,
  col.x,
  col.y,
  col.time,
  col.lead.time,
  time.series,
  col.value,
  verbose = T,
  preview = 2,
  remove.duplicates = T,
  circular.pars = NULL,
  col.station.name = NULL,
  show.progress = F,
  sort.stations = NULL
)
}
\arguments{
\item{df}{A data frame to be converted to an R list.}

\item{col.par}{The column name for parameter names.}

\item{col.x}{The column name for station x coordinates.}

\item{col.y}{The column name for station y coordinates.}

\item{col.time}{The column name for forecast cycle time. The column should be POSIXct.}

\item{col.lead.time}{The column name for forecast lead time. The column should be numeric
for the number of seconds from the forecast cycle time.}

\item{time.series}{The forecast cycle times to be extract into AnEn Forecasts. This should
be a POSIXct vector.}

\item{col.value}{The column name for data values.}

\item{verbose}{Whether to print progress messages.}

\item{preview}{How many entries to preview in progress messages.}

\item{remove.duplicates}{Whether to remove redundant values associated
with the same time. Sometimes, it is possible that, due to equipment mulfunctions,
there are multiple measurements at the same time. Idealy this should cleaned prior
to this function, but this function is able to keep the first appearance and remove
the rest.}

\item{circular.pars}{A character vector for the circular parameter names.}

\item{col.station.name}{The column name for station names.}

\item{show.progress}{Whether to show a progress bar.}

\item{sort.stations}{Sort station. It can be \code{Xs}, \code{Ys}, or \code{StationNames} if it is set.}
}
\value{
An R list for AnEn Forecasts.
}
\description{
RAnEn::formatForecasts generates the AnEn Forecasts list required
by analog computation.
}
\details{
The AnEn Forecasts is an R list with members including \code{ParameterNames},
\code{Xs}, \code{Ys}, \code{Data}, and etc., with a full list accessible
\href{https://analogensemble.ddns.net/2019/01/16/NetCDF-File-Types.html#forecasts}{here}.
RAnEn::formatForecasts make it easier to convert a data frame into
a such list data structure.

I read \href{https://cran.r-project.org/web/packages/dplyr/vignettes/programming.html}{this tutorial}
when developing functions with \code{dplyr}. It is very informative of using variables
with \code{dplyr} functions.
}
\seealso{
RAnEn::formatObservations
}
