% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generateAnalogs.default}
\alias{generateAnalogs.default}
\title{RAnEn::generateAnalogs}
\usage{
\method{generateAnalogs}{default}(sx_forecasts, sx_observations,
  sx_test_times, sx_search_times, sx_config = NULL,
  sx_algorithm = NULL)
}
\arguments{
\item{sx_forecasts}{Forecasts. You can generate this by typing
\code{forecasts <- generateForecastsTemplate()}.}

\item{sx_observations}{Observations. You can generate this by typing
\code{observations <- generateObservationsTemplate()}}

\item{sx_test_times}{Test times}

\item{sx_search_times}{Search times}

\item{sx_config}{A Config. You can generate this by typing \code{config <- new(Config)}}

\item{sx_algorithm}{\code{IS} for independent search or \code{SSE} for search
space extension}
}
\value{
A list of results with class AnEn. You can type \code{print(AnEn)}
to see a summary.
}
\description{
RAnEn::generateAnalogs generates analog ensemble.
}
