% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCircleCoords.R
\name{generateCircleCoords}
\alias{generateCircleCoords}
\title{RAnEn::generateCircleCoords}
\usage{
generateCircleCoords(x, y, radius, np = 100)
}
\arguments{
\item{x}{x coordinate of the center.}

\item{y}{y coordinate of the center.}

\item{radius}{radius of the circle in user's unit.}

\item{np}{number of points to generate on the circle.
By default it is set to 100.}
}
\value{
a matrix with a dimension of np x 2.
}
\description{
Generate a number of sequential point coordinates that
lie on a circle based on the center coordinates of the
circle and the radius.
}
\examples{
pts <- generateCircleCoords(3, 5, 2)
plot(pts, pch = 19, cex = 0.5, asp = 1)
points(3, 5, pch = 19, col = 'red')

}
\author{
Weiming Hu, \email{weiming@psu.edu}
}
