% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateConfiguration.R
\name{generateConfiguration}
\alias{generateConfiguration}
\title{RAnEn::generateConfiguration}
\usage{
generateConfiguration(mode, advanced = F)
}
\arguments{
\item{mode}{A character string of either 'independentSearch' or
'extendedSearch'.}

\item{advanced}{Whether to generate advanced configuration.}
}
\value{
A configuration list with all the parameters you need to specify.
Please don't change the mode of the configuration mode. If you need a configuration
with a different mode, you need to call the function again with the desired mode.
}
\description{
RAnEn::generateConfiguration generates a templated configuration object to
be passed to the function \code{\link{generateAnalogs}}.
This object can be validated before use by the function
\code{\link{validateConfiguration}}.
}
\details{
For configuration with mode 'independentSearch':
\itemize{
\item mode: 'independentSearch' indicates only searching the forecasts at the current location.
\item forecasts: The 4-dimensional forecast array with dimensions (parameters, stations, days, forecast lead times).
\item forecast_times: Times for forecasts. This should have the same length as the third dimension of forecasts.
\item flts: FLT information for test and search forecasts.
\item search_observations: The observations from which analog ensemble members are selected.
\item observation_times: Time information for search observations.
\item observation_id: The parameter index for which the observation parameter to use. For example,
if you have several observation parameters in search observations, this parameter tells the
package which observation parameter to use.
\item num_members: The number of members to keep in the ensembles of analogs.
\item circulars: The index of forecast parameters in search forecasts that are circular.
\item weights: The weight for each parameter in forecasts.
\item quick: Whether to use quick sort algorithm.
\item extend_observations: TRUE indicates observations should be taken from extended search stations,
FALSE indicates observations should be taken from the current station.
\item preserve_similarity: Whether to preserve the similarity matrix in returned AnEn.
\item preserve_mapping: Whether to preserve the mapping matrix in return AnEn.
\item preserve_search_stations: Whether to preserve the search station list in AnEn.
\item preserve_std: Whether to preserve the standard deviation for search forecasts. The results might not
not be correct if operational search is used because the standard deviation will be changed due to the
increase of search days. But the changes are not reflected in the R results.
\item verbose: Verbose level. You can find settings for different values in the
\href{https://weiming-hu.github.io/AnalogsEnsemble/CXX/class_an_en.html#a25984b953516a987e2e9eb23048e5d60}{C++ documentation verbose}.
\item debug: Whether to return debug flags when NA values present. These debug flags will be negative values for NA indices that
indicate the cause for this NA values. Please see the C++ functions
\href{https://weiming-hu.github.io/AnalogsEnsemble/CXX/class_an_en.html#a787e43e7bfdaa830bf70fa6710731614}{computeSimilarity} and
\href{https://weiming-hu.github.io/AnalogsEnsemble/CXX/class_an_en.html#a8e2b88cda5cc9fce8ea4703a3236719c}{selectAnalogs}.
\item time_match_mode: This guides how to compute the time mapping from forecast times and forecast
lead times to observation times. 0 for strict search (return error when matching observation times
cannot be found) and 1 for loose search.
\item max_par_nan: The number of NAN values allowed when averaging similarity across different parameters.
Set it to a negative value to allow any number of NAN values.
\item max_flt_nan: The number of NAN values allowed when averaging FLT window similarity.
Set it to a negative value to allow any number of NAN values.
\item test_times_compare: The times in test forecast times that will be compared and
similarity will be generated for them. This time should be selected from the test_times field.
\item search_times_compare: The times in search times that will be compared and similarity will be generated from them.
This time should be selected from the search_times field.
\item operational: Whether to use operational serach mode and increase the number
of search times while moving forward with test times.
\strong{Operational serach times are always generated from the search_times.}
\item max_num_sims: The maximum number of similarity to keep in the array.
This can be used to reduce the memory requirement. By default, twice the number
of analog members are kept in the similarity record. Set this to a negative value to keep all records.
\item FLT_radius: The radius for comparing a parameter along lead times.
This specifies how many prior and subsequent lead times to compare. Must be non-negative.
}

For configuration with mode 'extendedSearch': (skipping the aforementioned parameters)
\itemize{
\item mode: 'extendedSearch' searches forecasts from both the current and nearby locations.
\item forecast_stations_x: X coordinates of forecast stations/grid points.
\item forecast_stations_y: Y coordinates of forecast stations/grid points.
\item stations_classifier: The classifier (tag) for each test stations.
\item num_nearest: The number of nearest neighbors.
\item distance: The distance to search for nearby stations/grid points.
\item max_num_search_stations: The maximum search stations to keep in the similarity matrix. The use of this parameter
will only affect the number of search stations to keep in the similarity matrix. This won't affect the correctness
of nearby station selection. This parameter is used in case of insufficient memeory.
If \code{distance} is the only constraint of including nearby stations, to avoid including too many stations within
the specified distance, this should be set.
}

Please find detailed terminology explanation in the
\href{https://weiming-hu.github.io/AnalogsEnsemble/2018/10/06/vocabulary.html}{Vocabulary post}.
If you still can't find helpful information or can't understand the parameters, please open a ticket.
}
\author{
Weiming Hu \email{weiming@psu.edu}
}
