% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateTemplates.R
\name{generateForecastsTemplate}
\alias{generateForecastsTemplate}
\title{RAnEn::generateForecastsTemplate}
\usage{
generateForecastsTemplate()
}
\value{
An empty list with members
}
\description{
RAnEn::generateForecastsTemplate generates an empty forecasts.
This will guide you to create the correct forecasts that can
then be used in RAnEn::generateAnalogs.
}
