% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateTemplates.R
\name{generateObservationsTemplate}
\alias{generateObservationsTemplate}
\title{RAnEn::generateObservationsTemplate}
\usage{
generateObservationsTemplate()
}
\value{
An empty list with members
}
\description{
RAnEn::generateObservationsTemplate generates an empty observations.
This will guide you to create the correct observations. that can
then be used in RAnEn::generateAnalogs.
}
