% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePersistent.R
\name{generatePersistent}
\alias{generatePersistent}
\title{RAnEn::generatePersistent}
\usage{
generatePersistent(config, forecast.time.interval = 86400,
  show.progress = T, silent = F)
}
\arguments{
\item{config}{A Configuration object created from \code{\link{generateConfiguration}}.}

\item{forecast.time.interval}{The forecast time interval in seconds. This is
usually \code{24 * 60 * 60} because the third dimension of forecasts is usually day.
If it is half day, change the value accordingly.}

\item{show.progress}{Whether to show a progress bar.}

\item{silent}{Parameter passed to \code{\link{alignObservations}}.}
}
\value{
An 5-dimension array. The format is exactly the same as the member
\code{analogs} specified in \code{\link{generateAnalogs}}.
}
\description{
RAnEn::generatePersistent generates persistent analog ensemble forecasts.
For each test forecast, it simply takes the historical observation values
that are associated with the historical forecasts that are directly prior
to the current forecasts. The number of values taken would be the number of
ensemble members desired.
}
\details{
For example, to generate a 3-member persistent analog ensemble for a particular
forecast at a certain grid point, a certain lead time, and on July 25, 2019,
the \strong{observations} associated with the historical forecasts on July 22,
24, and 24 will be included in the persistent analogs.
}
\author{
Weiming Hu \email{weiming@psu.edu}
}
