% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generateSearchStations}
\alias{generateSearchStations}
\title{RAnEn::generateSearchStations}
\usage{
generateSearchStations(sx_xs, sx_ys, num_neighbors, sx_distance = NULL)
}
\arguments{
\item{sx_xs}{X coordinates}

\item{sx_ys}{Y coordinates}

\item{num_neighbors}{The number of neighbors to find}

\item{distance}{The distance threshold when finding nearest neighbors}
}
\value{
A matrix for nearest search station index. Each row is a set of search station index
}
\description{
RAnEn::generateSearchStations calculates the nearest coordinates for all input coordinates.
}
