% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generateTimeMapping}
\alias{generateTimeMapping}
\title{RAnEn::generateTimeMapping}
\usage{
generateTimeMapping(sx_fcst_times, sx_fcst_flts, sx_obs_times)
}
\arguments{
\item{sx_fcst_times}{forecast times}

\item{sx_fcst_flts}{forecast lead times}

\item{sx_obs_times}{observation times}
}
\value{
A matrix with rows being forecast times and columns being lead times, the cell
values are the index to the corresponding observation times.
}
\description{
RAnEn::generateTimeMapping generates the index mapping table from foreacst times and lead times
to observation times.
}
