% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAssociatedForecasts.R
\name{getAssociatedForecasts}
\alias{getAssociatedForecasts}
\title{RAnEn::getAssociatedForecasts}
\usage{
getAssociatedForecasts(
  AnEn,
  config,
  forecast.parameter,
  i.station = NA,
  i.test.day = NA,
  i.flt = NA,
  similarity.name = "similarity",
  forecasts.name = "forecasts",
  members.to.extract = NA
)
}
\arguments{
\item{AnEn}{The AnEn class object returned by \code{\link{generateAnalogs}}.}

\item{config}{The Configuration class object used to generate AnEn.}

\item{forecast.parameter}{The parameter index for which variable to extract
in the forecast array in \code{config}.}

\item{i.station}{The station index in AnEn to extract.}

\item{i.test.day}{The test day index in AnEn to extract.}

\item{i.flt}{The FLT index in AnEn to extract.}

\item{similarity.name}{The similarity name to be used in \code{AnEn}. This
is useful when you have a different similarity matrice name to the regular one.}

\item{forecasts.name}{The forecast name to be used in \code{config}. This
is useful when you have a different similarity matrice name to the regular one.}

\item{members.to.extract}{The number of ensemble members to extract. By default
to extract all members.}
}
\value{
An array with the same dimension as analogs in AnEn. For more
information of each dimension, please refer to \code{\link{generateAnalogs}}.
The only difference is that the values in this array are extracted
from the most similar past forecast that is associated with each analog ensemble
member.
}
\description{
RAnEn::getAssociatedForecasts returns the most similar
past forecasts that are identified using this analog ensemble
and associated with each analog ensemble member. In
other words, this funtion returns the most similar
past forecasts to the current forecasts.
}
\author{
Weiming Hu \email{weiming@psu.edu}

Laura Clemente-Harding \email{laura@psu.edu}
}
