% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTransparent.R
\name{makeTransparent}
\alias{makeTransparent}
\title{RAnEn::makeTransparent}
\usage{
makeTransparent(col = stop("Please provide a vector of colours."),
  alpha = 1)
}
\arguments{
\item{col}{The color(s) to be made transparent.}

\item{alpha}{The transparency value.}
}
\description{
RAnEn::makeTransparent adds the transparent parameter `alpha` to a color.
}
\author{
Guido Cervone \email{cervone@psu.edu}

Martina Calovi \email{mxc895@psu.edu}

Laura Clemente-Harding \email{laura@psu.edu}
}
