% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeList.R
\name{mergeList}
\alias{mergeList}
\title{RAnEn::mergeList}
\usage{
mergeList(from, to, names)
}
\arguments{
\item{from}{A list to copy from}

\item{to}{A list to copy to}

\item{names}{the names to copy, usually \code{names(from)}}
}
\description{
RAnEn::mergeList combine two lists. It does not allow names in conflict
unless the name is further a list. This is similar to a depth-first copy
for list variables.
}
\examples{
library(RAnEn)
from <- list(C = list(C = 3, D = list(A = 1)), D = 3)
to <- list(A = 1, B = 2, C = list(A = 1, B = 2))
merged <- mergeList(from, to, names(from))

print(from, recursive = T)
print(to, recursive = T)
print(merged, recursive = T)

}
\author{
Weiming Hu \email{weiming@psu.edu}
}
