% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncvar_get_times.R
\name{ncvar_get_times}
\alias{ncvar_get_times}
\title{RAnEn::ncvar_get_times}
\usage{
ncvar_get_times(
  nc,
  var_name,
  origin = "1970-01-01",
  tz = "UTC",
  to_posix = T,
  epsilon = 1e-05
)
}
\arguments{
\item{nc}{An object of class ncdf4 (as returned by either function nc_open or function nc_create), indicating what file to read from.}

\item{var_name}{The variable name for \code{Times}}

\item{origin}{The origin used for converting to POSIX time}

\item{tz}{The time zone}

\item{to_posix}{Whether to convert times to POSIX}

\item{epsilon}{The maximum reading error to correct. Otherwise, an error is thrown.}
}
\value{
A vector of times
}
\description{
RAnEn::ncvar_get_times is a convenient function to read Times from
a already open NetCDF file connection. It echos the API of \code{ncdf4:ncvar_get}
but it differs in that it checks for any reading imprecision and corrects
for small errors.
}
\author{
Weiming Hu \email{weiming@psu.edu}
}
