% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAnalogSample.R
\name{plotAnalogSample}
\alias{plotAnalogSample}
\title{RAnEn::plotAnalogSample}
\usage{
plotAnalogSample(AnEn, config, i.station, i.test.day, obs.id, fcst.id,
  parameter.names, share.parameter.names = F, member.name = "analogs",
  flts.fraction = 3600, boxplot.outliers = F,
  border.anen.box = "grey", lty.current.fcsts = "longdash",
  col.current.fcsts = "red", cex.current.fcst = 1,
  pch.current.obs = 1, col.current.obs = "black",
  cex.current.obs = 1.5, legend.texts = c("Observations", "NAM"),
  legend.position = "topleft", single.figure = T, mar = c(1.5, 5, 1,
  1) + 0.1, omi = c(0.3, 0, 0, 0))
}
\arguments{
\item{AnEn}{An AnEn object generated from \code{\link{generateAnalogs}}.}

\item{config}{A Configuration object generated from \code{\link{generateConfiguration}}.}

\item{i.station}{The station index from analogs.}

\item{i.test.day}{The test day index from analogs.}

\item{obs.id}{The index in observation parameters that which parameter is predicted.}

\item{fcst.id}{The index in forecast parameters that which parameter is predicted.}

\item{parameter.names}{The forecast parameter names.}

\item{share.parameter.names}{Whether forecasts and observations have the same
parameters. If this is TRUE, the observations will be plot on the other parameters
that are used to compute the simialrity.}

\item{member.name}{Which member from AnEn should be plotted. This is useful
when you have bias corrected analog members in AnEn that you would like to plot.}

\item{flts.fraction}{This is the fraction when dividing FLT values. 3600 means that
the x labels are in hours when FLTs are in seconds.}

\item{boxplot.outliers}{Whether to plot outliers.}

\item{border.anen.box}{The color of the box plots.}

\item{lty.current.fcsts}{The line type for current forecasts.}

\item{col.current.fcsts}{The color for current forecasts.}

\item{cex.current.fcst}{The size for current forecasts.}

\item{pch.current.obs}{The point type for current observations.}

\item{col.current.obs}{The color for current observations.}

\item{cex.current.obs}{The size for current observations.}

\item{legend.texts}{A vector of characters of length 2. The first one
is for observations, and the second one for forecasts.}

\item{legend.position}{The legend position.}

\item{single.figure}{Whether to generate a single plot with subplots.}

\item{mar}{Margin for the plot. This is effective only when single.figure is TURE.}

\item{omi}{Outer margin for the plot. This is effective only when single.figure is TURE.}
}
\description{
RAnEn::plotAnalogSample compares the current forecasts and AnEn with observations
for one particular station and test time across all FLTs available in forecasts.
It also plots the historical forecasts that are identified by AnEn, and
compares them with current forecasts for each parameters used to generate
analogs.
}
\author{
Weiming Hu \email{cervone@psu.edu}

Laura CLemente-Harding \email{lec170@psu.edu}

Martina Calovi \email{mcalovi@psu.edu}

Guido Cervone \email{cervone@psu.edu}
}
