% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAnalogSelection.R
\name{plotAnalogSelection}
\alias{plotAnalogSelection}
\title{RAnEn::plotAnalogSelection}
\usage{
plotAnalogSelection(AnEn, config, i.station, i.test.day, i.flt,
  parameter.names, cex.lab = 1.5, pch.selected = 16,
  pch.unselected = 16, col.selected = "red",
  col.unselected = "lightgrey", col.reference = "black",
  lty.reference = "longdash", lwd.reference = 2,
  origin = "1970-01-01", tz = "UTC", single.figure = T,
  mar = c(1.5, 5, 1, 1) + 0.1, omi = c(0.3, 0, 0, 0), use.plotly = F)
}
\arguments{
\item{AnEn}{An AnEn object generated from \code{\link{generateAnalogs}}.}

\item{config}{A Configuration object generated from \code{\link{generateConfiguration}}.}

\item{i.station}{The station index from analogs.}

\item{i.test.day}{The test day index from analogs.}

\item{i.flt}{The FLT index from analogs.}

\item{parameter.names}{The parameter names that are associated with forecasts in the configuration.}

\item{cex.lab}{The font size of labels.}

\item{pch.selected}{The point type for selected points.}

\item{pch.unselected}{The point type for unselected points.}

\item{col.selected}{The color for selected points.}

\item{col.unselected}{The color for unselected points.}

\item{col.reference}{The color for the reference line.}

\item{lty.reference}{The line type for the reference line.}

\item{lwd.reference}{The line width for the reference line.}

\item{origin}{The origin for times in forecasts. This will be passed to as.POSIXct to convert 
times in forecasts to date/time objects.}

\item{tz}{The time zone in forecasts. This will be passed to as.POSIXct to convert 
times in forecasts to date/time objects.}

\item{single.figure}{Whether to generate only a single figure.}

\item{mar}{Margins for subplots. This is effective only when single.figure is TURE.
It is passed to the function par.}

\item{omi}{Outer margin for the plot. This is effective only when single.figure is TURE.
It is passed to the function par.}
}
\description{
RAnEn::plotAnalogSelection visualize the selection of analog ensemble members.
It shows how ensemble members are selected based on the similarity metrics.
It also shows the selected forecasts for each parameter used during the 
analog generation, compared with the other unselected forecasts.
}
\author{
Weiming Hu \email{cervone@psu.edu}

Laura CLemente-Harding \email{lec170@psu.edu}

Martina Calovi \email{mcalovi@psu.edu}

Guido Cervone \email{cervone@psu.edu}
}
