% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAnEnMap.R
\name{plotAnEnMap}
\alias{plotAnEnMap}
\title{RAnEn::plotAnEnMap}
\usage{
plotAnEnMap(ras, z = "obs", spdf = NULL, spdf.bg = NULL,
  spdf.coastline = NULL, col = rainbow(9),
  col.coastline = "darkgrey", col.bg = "white", col.text = "black",
  col.interpolate = T, border = "darkgrey", xlab = "Longitude",
  ylab = "Latitude", lwd = 1, cex = 1, lwd.coastline = 1,
  lwd.bg = 1, smallplot = c(0.1, 0.14, 0.6, 0.9), legend = "",
  legend.line = 1, legend.adj = 0, add.legend = T, zlim = NULL,
  ext = NULL)
}
\description{
RAnEn::plotAnEnMap documentation underdevelopment.
}
\author{
Guido Cervone \email{cervone@psu.edu}
}
