% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAnEnTimeSeries.R
\name{plotAnEnTimesSeries}
\alias{plotAnEnTimesSeries}
\title{RAnEn::plotAnEnTimesSeries}
\usage{
plotAnEnTimesSeries(analogs, mod, obs, col.frame = "grey",
  col.bg = "lightgrey", col.lines = "bisque4", ylim = NULL, n = 10)
}
\description{
RAnEn::plotAnEnTimesSeries documentation underdevelopment.
}
\author{
Guido Cervone \email{cervone@psu.edu}
}
