% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBackground.R
\name{plotBackground}
\alias{plotBackground}
\title{RAnEn::plotBackground}
\usage{
plotBackground(col = "lightgrey")
}
\description{
RAnEn::plotBackground is a sugar function to plot a background color.
}
\author{
Guido Cervone \email{cervone@psu.edu}
}
