% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Configuration.R
\name{print.Configuration}
\alias{print.Configuration}
\title{RAnEn::print.Configuration}
\usage{
\method{print}{Configuration}(x, forecasts.names = NA)
}
\arguments{
\item{x}{a Configuration object}

\item{forecasts.names}{The name of each forecast variable (the first dimension in the forecast array).}
}
\description{
Overload print function for class Configuration
}
\examples{
config <- generateConfiguration('independentSearch')
print(config)

}
\author{
Weiming Hu \email{weiming@psu.edu}
}
