% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.list.R
\name{print.list}
\alias{print.list}
\title{RAnEn::print.list}
\usage{
\method{print}{list}(x)
}
\arguments{
\item{x}{The object to print}

\item{names_to_print}{The names to print. Default to all names.}
}
\description{
RAnEn::print.list prints a list in pretty format
}
\author{
Weiming Hu \email{weiming@psu.edu}
}
