% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Templates.R
\name{print.Observations}
\alias{print.Observations}
\title{RAnEn::print.Observations}
\usage{
\method{print}{Observations}(x)
}
\arguments{
\item{x}{A Forecasts object}
}
\description{
RAnEn::print.Observations print a Forecasts object.
}
