% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAnEn.R
\docType{package}
\name{RAnEn-package}
\alias{RAnEn}
\alias{RAnEn-package}
\title{RAnEn::RAnEn}
\description{
The Analog Ensemble Package computes analog ensembles. It is implemented in C++ and exported to R interface.
}
\details{
RAnEn is the R package implementing the Analog Ensemble method and more with data processing and visualization tasks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://analogensemble.ddns.net/}
  \item \url{https://github.com/uga-gaim/AnalogEnsemble}
  \item Report bugs at \url{https://github.com/uga-gaim/AnalogEnsemble/issues}
}

}
\author{
Weiming Hu \email{weiming@psu.edu}

Guido Cervone \email{cervone@psu.edu}

Martina Calovi \email{mxc895@psu.edu}

Laura Clemente-Harding \email{laura@psu.edu}
}
\keyword{Analog}
\keyword{Ensemble,}
\keyword{Forecasts}
\keyword{Probabilistic}
\keyword{Weather}
