% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAnEn.R
\name{readAnEn}
\alias{readAnEn}
\title{RAnEn::readAnEn}
\usage{
readAnEn(file.analogs, file.similarity = NA, verbose = F)
}
\arguments{
\item{file.analogs}{The path to analog NetCDF file.}

\item{file.similarity}{The path to similarity NetCDF file.}

\item{verbose}{Whether to print progress.}
}
\value{
An AnEn class object.
}
\description{
RAnEn::readAnEn can read AnEn results from NetCDF files that are generated 
from analogGenerator. It supports reading analogs and similarity results.
}
\author{
Weiming Hu \email{weiming@psu.edu}
}
