% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readConfig.R
\name{readConfig}
\alias{readConfig}
\title{RAnEn::readConfig}
\usage{
readConfig(config.file, replace.dash = "_")
}
\arguments{
\item{config.file}{The full path to the configuration file (.cfg)}

\item{replace.dash}{A single character to replace the dash in configuration names for. This can be set to \code{NULL}
if no replacement is desired.}
}
\value{
A list
}
\description{
RAnEn::readConfig reads a config file, usually generated by RAnEn::writeConfig
}
\author{
Weiming Hu \email{weiming@psu.edu}
}
