% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNcAttrs.R
\name{readNcAttrs}
\alias{readNcAttrs}
\title{RAnEn::readNcAttrs}
\usage{
readNcAttrs(file, recursive)
}
\arguments{
\item{file}{The file to read}

\item{recursive}{Whether to read attributes recursively}
}
\description{
RAnEn::readNcAttrs reads NetCDF attributes. Only attributes under
groups or in the root group are read. Attributes associated with variables
are not read.
}
\author{
Weiming Hu \email{weiming@psu.edu}
}
