% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readObservations.R
\name{readObservations}
\alias{readObservations}
\title{RAnEn::readObservations}
\usage{
readObservations(file, origin = "1970-01-01", tz = "UTC")
}
\arguments{
\item{file}{The observation NetCDF file to read.}

\item{origin}{The origin of time to be used in as.POSIXct.}

\item{tz}{The time zone of time to be used in as.POSIXct.}
}
\value{
A list object.
}
\description{
RAnEn::readObservations reads observation data from the NetCDF file with a
\href{https://analogensemble.ddns.net/2019/01/16/NetCDF-File-Types.html}{standard format}.
Most often, the file should be created from gribConverter.
}
\author{
Weiming Hu \email{weiming@psu.edu}
}
