% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readOptional.R
\name{readOptional}
\alias{readOptional}
\title{RAnEn::readOptional}
\usage{
readOptional(obj, names, nc, group_prefix)
}
\arguments{
\item{names}{The optional names to read}

\item{nc}{An NetCDF file object}

\item{group_prefix}{The group prefix to use to find the variables}

\item{obs}{Observations or Forecasts}
}
\value{
The input Observations or Forecasts but with extra names
}
\description{
RAnEn::readOptional reads the input optional names from an NetCDF file object
with \code{read} file mode.
}
\author{
Weiming Hu \email{weiming@psu.edu}
}
\keyword{internal}
