% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showVerificationMethods.R
\name{showVerificationMethods}
\alias{showVerificationMethods}
\title{RAnEn::showVerificationMethods}
\usage{
showVerificationMethods()
}
\description{
RAnEn::showVerificationMethods prints all the supported verification methods.
}
