% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortStations.R
\name{sortStations}
\alias{sortStations}
\title{RAnEn::sortStations}
\usage{
sortStations(obj, sort.stations = NULL, verbose = F)
}
\arguments{
\item{obj}{A Forecasts or Observations}

\item{sort.stations}{Either one of \code{Xs}, \code{Ys}, or \code{StationNames}, or a numeric
vector for the order. For example, if a numeric vector is passed, like \code{c(1, 4, 2, 3)},
the stations and the data will be ordered in such a way that: the first station is not
moved, the second station is moved to the third position, the third station is moved
to the forth position, and the forth station is moved to the second position. And
there are only four stations!}
}
\description{
RAnEn::sortStations sorts the data based on either 'Xs', 'Ys', or 'StationNames'.
}
