% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetCoordinates.R
\name{subsetCoordinates}
\alias{subsetCoordinates}
\title{RAnEn::subsetCoordinates}
\usage{
subsetCoordinates(xs, ys, poi, file.output = NULL,
  arg.name = stop("Specify the argument name!"))
}
\arguments{
\item{xs}{A vector with X coordinates.}

\item{ys}{A vector with Y coordinates.}

\item{poi}{Points of interest, or target points. It can be either a named
vector with \code{left}, \code{right}, \code{top}, and \code{bottom} or a data frame for
a set of target point coordinates as rows with variable names \code{X} and \code{Y}.}

\item{file.output}{A file name to write the ID of subset stations.}

\item{arg.name}{If you are going to use the output file with \code{fileSlice},
this is \code{slice-stations}; if the app is \code{gribConverter}, this is \code{subset-stations}.}
}
\value{
A data frame with ID and coordinates for subset stations. The column \code{ID.C}
should be used for C++ programs. The column \code{ID.R} should be used in R.
}
\description{
RAnEn::subsetCoordinates extracts the coordinates within an predefined extent
or the coordinates that are closest to a set of target coordinates based
on Eucleadian distances. This function can also write the corresponding
IDs to the extracted coordinates into a file so that it can be used
in \href{https://github.com/Weiming-Hu/AnalogsEnsemble/tree/master/apps}{other executables provided}
in the PAnEn packages.
}
\examples{

####################################################################
#                       Use Case 1                                 #
####################################################################

# Define the forecast grid coordinates
forecast.grids <- expand.grid(X = 1:10, Y = 1:10)

# Define the observation coordinates
obs.coords <- data.frame(X = c(3.2, 7.9), Y = c(5.1, 4.2))

# Find the closest forecast grid to each observation location
subset.coords <- subsetCoordinates(
  forecast.grids$X, forecast.grids$Y, obs.coords)

# Visualization
plot(forecast.grids$X, forecast.grids$Y, pch = 16,
     cex = 1, col = 'grey', asp = 1, xlab = '', ylab = '')
points(obs.coords$X, obs.coords$Y, pch = 16, cex = 1, col = 'red')
points(subset.coords$X, subset.coords$Y, cex = 1.5, lwd = 1.5)
legend('top', legend = c('Forecast Grid', 'Observation', 'Cloest Grid'),
       col = c('grey', 'red', 'black'), pch = c(16, 16, 1),
       horiz = T, cex = 1.2)

####################################################################
#                       Use Case 2                                 #
####################################################################

# Define a region of interest
extent <- c('left' = 2.5, 'right' = 8,
            'top' = 8.9, 'bottom' = 2)

# Subset the forecast grids within the extent
subset.coords <- subsetCoordinates(
  forecast.grids$X, forecast.grids$Y, extent)

# Visualization
plot(forecast.grids$X, forecast.grids$Y, pch = 16,
     cex = 1, col = 'grey', asp = 1, xlab = '', ylab = '')
rect(xleft = extent['left'], xright = extent['right'],
     ybottom = extent['bottom'], ytop = extent['top'],
     cex = 1, col = NA, lwd = 1, lty = 'dashed')
points(subset.coords$X, subset.coords$Y, cex = 1.5, lwd = 1.5)
legend('top', legend = c('Forecast Grid', 'Subset Grid'),
       col = c('grey', 'black'), pch = c(16, 1),
       horiz = T, cex = 1.2)

}
\author{
Weiming Hu \email{weiming@psu.edu}
}
