% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetStations.R
\name{subsetStations}
\alias{subsetStations}
\title{RAnEn::subsetStations}
\usage{
subsetStations(index, l, verbose = T)
}
\arguments{
\item{index}{An index vector for stations to extract}

\item{l}{A forecast or observation list. For how to create such a
list, please see
\href{https://weiming-hu.github.io/AnalogsEnsemble/2019/11/18/format-obs.html}{this tutorial}.
For what members to include in the list, see
\href{https://weiming-hu.github.io/AnalogsEnsemble/2019/01/16/NetCDF-File-Types.html}{this doc}.}

\item{verbose}{Whether to print progress information.}
}
\value{
A forecast or observation list depending on your input
list type with the subset stations.
}
\description{
RAnEn::subsetStations is a convevient function to subset stations
from forecast and observation lists. Given an index vector, presumably
from the \code{ID.R} column from RAnEn::subsetCoordinates, this function
can subset the desired stations from a forecast or observation list.
}
\details{
RAnEn::subsetStations will select the stations based on the input
index from the following members of the input list (if they exist):
\itemize{
\item StationNames
\item Xs
\item Ys
\item Data
}
}
\examples{
# Subset a forecast list
forecasts <- list(
	StationNames = paste0('Station', 1:10),
	Xs = 1:10, Ys = 10:1,
	Data = array(1:200, dim = c(2, 10, 5, 2))
)

index <- c(1, 5)

forecasts.sub <- subsetStations(index, forecasts)

forecasts.sub$StationNames

dim(forecasts$Data)

dim(forecasts.sub$Data)

# Subset an observation list
observations <- list(
	StationNames = paste0('Station', 1:10),
	Xs = 1:10, Ys = 10:1,
	Data = array(1:100, dim = c(2, 10, 5))
)

index <- 6

observations.sub <- subsetStations(index, observations)

observations.sub$StationNames

dim(observations$Data)

dim(observations.sub$Data)


}
