% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toDateTime.R
\name{toDate}
\alias{toDate}
\title{RAnEn::toDate}
\usage{
toDate(time.in.secs, origin = "1970-01-01", tz = "UTC")
}
\description{
RAnEn::toDate is a sugar function to convert seconds to date time objects.
}
\author{
Guido Cervone \email{cervone@psu.edu}
}
