% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toForecastTime.R
\name{toForecastTime}
\alias{toForecastTime}
\title{RAnEn::toForecastTime}
\usage{
toForecastTime(observation.time, mapping, flt = NA)
}
\arguments{
\item{observation.time}{The observation time index(es) from the \code{Time} column of the \code{analogs} member in
in the results of \code{\link{generateAnalogs}}.}

\item{mapping}{The mapping table from the \code{mapping} member in the results of \code{\link{generateAnalogs}}.
Or it can be created from \code{\link{generateTimeMapping}}.}

\item{flt}{The lead time index.}
}
\description{
RAnEn::toForecastTime converts the observation time index to forecast time index.
}
\details{
This function does the following conversion.

\verb{Observation time index [ + Forecast lead time index ] ----> Forecast time index}

To pin down the exact time of a forecast, two things are needed: forecast time and the lead time; to pin down
the exact time of an observation, only the observation time is needed. Therefore, this function is devoted to
convert from observation times to forecast times. Please note that, an observation time \strong{CAN} be associated
with multiple forecast times because of the existence of forecast lead times. If you have constraints on the
lead time, then the match between forecasts and observations becomes one-on-one.

The observation time index usually comes from the \code{Time} column of the \code{analogs} member in the results of
\code{\link{generateAnalogs}}. This time is observation times. However, please note that the times in the
\code{similarity} member are \strong{forecast times}. These two are different.
}
\author{
Guido Cervone \email{cervone@psu.edu}

Martina Calovi \email{mxc895@psu.edu}

Laura Clemente-Harding \email{laura@psu.edu}
}
