% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toMemberForecasts.R
\name{toMemberForecasts}
\alias{toMemberForecasts}
\title{RAnEn::toMemberForecasts}
\usage{
toMemberForecasts(forecasts, sims.time.index, sims.stations.index = NULL,
  stations.index = 1, test.times.index = 1, flts.index = 1,
  members.index = NULL, parameters.index = NULL, verbose = T)
}
\arguments{
\item{forecasts}{The Data from Forecasts. Usually it is a 4D array.}

\item{sims.time.index}{The similarity time index array from AnEn.}

\item{sims.stations.index}{The similarity station index array from AnEn.}

\item{stations.index}{The station index of the analogs.}

\item{test.times.index}{The test time index of the analogs.}

\item{flts.index}{The forecast lead time index of the analogs.}

\item{members.index}{The analog members index.}

\item{parameters.index}{The forecast parameters index.}

\item{verbose}{Whether to show progress and summary information}
}
\value{
A 5 dimensional array with the shape \code{[parameters][stations][times][flts][members]}.
}
\description{
RAnEn::toMemberForecasts finds the most similar historical forecasts
that are associated with each analog members.
}
\author{
Weiming Hu \email{weiming@psu.edu}

Laura Clemente-Harding \email{laura@psu.edu}
}
