% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateConfiguration.R
\name{validateConfiguration}
\alias{validateConfiguration}
\title{RAnEn::validateConfiguration}
\usage{
validateConfiguration(x)
}
\arguments{
\item{x}{A configuration object generated by the function
\code{\link{generateConfiguration}}.}
}
\value{
A boolean whether the configuration object is fully configured and
ready to be used in computation.
}
\description{
RAnEn::validateConfiguration validates the configuration object created by
the function \code{\link{generateConfiguration}}.
}
\author{
Weiming Hu \email{weiming@psu.edu}
}
