% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.R
\name{verify}
\alias{verify}
\title{RAnEn::verify}
\usage{
verify(methods, verbose = T, ...)
}
\arguments{
\item{methods}{A vector of verification methods that should be carried out. Please
use \code{\link{showVerificationMethods}} to check the supported methods.}

\item{verbose}{Whether to print detail information.}
}
\value{
A list of verification results.
}
\description{
RAnEn::verify is the caller function to call various verification methods.
This is convenient when you want to carry out multiple verification at
once.
}
