% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifyCRMSE.R
\name{verifyCRMSE}
\alias{verifyCRMSE}
\title{RAnEn::verifyCRMSE}
\usage{
verifyCRMSE(anen.ver, obs.ver, boot = F, R = 1000, na.rm = T, parallel = F)
}
\arguments{
\item{boot}{Whether to use bootstrap.}

\item{R}{The number of bootstrap replicates. Used by the function \code{boot::boot}.}

\item{na.rm}{Whether to remove NA values.}

\item{parallel}{Whether to use parallel processing.}
}
\description{
RAnEn::verfiyCRMSE computes the center RMSE.
}
\details{
RMSE ^ 2 = CRMSE ^ 2 + Bias ^ 2

To set the number of cores to use when parallel is used,
\code{options(mc.cores = 8)}.
}
