% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifyCRPS.R
\name{verifyCRPS}
\alias{verifyCRPS}
\title{RAnEn::verifyCRPS}
\usage{
verifyCRPS(anen.ver, obs.ver, boot = F, R = 1000, int.step = 150,
  na.rm = T)
}
\arguments{
\item{anen.ver}{A 4-dimensional array. This array is usually created from the `value` column of
the `analogs` member in the results of \code{\link{generateAnalogs}}. The dimensions should be
`[stations, times, lead times, members]`.}

\item{obs.ver}{A 3-dimensional array. The dimensions should be `[stations, times, lead times]`.
You can generate the array using \code{\link{alignObservations}}.}

\item{boot}{Whether to use bootstrap.}

\item{R}{The number of bootstrap replicates. Used by the function `boot::boot`.}

\item{int.step}{When boot is used, this is the integral length.}

\item{na.rm}{Whether to remove NA values.}
}
\description{
RAnEn::verifyCRPS calculates CRPS.
}
\author{
Guido Cervone \email{cervone@psu.edu}

Martina Calovi \email{mxc895@psu.edu}

Laura Clemente-Harding \email{laura@psu.edu}
}
