% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifyMAE.R
\name{verifyMAE}
\alias{verifyMAE}
\title{RAnEn::verifyMAE}
\usage{
verifyMAE(anen.ver, obs.ver, boot = F, R = 1000, na.rm = T,
  parallel = F)
}
\description{
Underdevelopment.
}
\details{
To set the number of cores to use when parallel is used,
\code{options(mc.cores = 8)}.
}
