% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifyRMSE.R
\name{verifyRMSE}
\alias{verifyRMSE}
\title{RAnEn::verifyRMSE}
\usage{
verifyRMSE(anen.ver, obs.ver, boot = F, R = 1000, na.rm = T, parallel = F)
}
\arguments{
\item{anen.ver}{A 4-dimensional array. This array is usually created from the \code{value} column of
the \code{analogs} member in the results of \code{\link{generateAnalogs}}. The dimensions should be
\verb{[stations, times, lead times, members]}.}

\item{obs.ver}{A 3-dimensional array. The dimensions should be \verb{[stations, times, lead times]}.
You can generate the array using \code{\link{alignObservations}}.}

\item{boot}{Whether to use bootstrap.}

\item{R}{The number of bootstrap replicates. Used by the function \code{boot::boot}.}

\item{na.rm}{Whether to remove NA values.}

\item{parallel}{Whether to turn on parallel processing.}
}
\description{
RAnEn::verifyRMSE calculates root mean square errors.
}
\details{
RMSE ^ 2 = CRMSE ^ 2 + Bias ^ 2
}
\author{
Guido Cervone \email{cervone@psu.edu}

Martina Calovi \email{mxc895@psu.edu}

Laura Clemente-Harding \email{laura@psu.edu}

To set the number of cores to use when parallel is used,
\code{options(mc.cores = 8)}.
}
