% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifySpread.R
\name{verifySpread}
\alias{verifySpread}
\title{RAnEn::verifySpread}
\usage{
verifySpread(anen.ver, na.rm = T, parallel = F)
}
\description{
Underdevelopment.
}
\details{
To set the number of cores to use when parallel is used,
\code{options(mc.cores = 8)}.
}
