% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifyThreatScore.R
\name{verifyThreatScore}
\alias{verifyThreatScore}
\title{RAnEn::verifyThreatScore}
\usage{
verifyThreatScore(anen.ver, obs.ver, threshold, ensemble.func, ...)
}
\arguments{
\item{anen.ver}{A 4-dimensional array for analogs.}

\item{obs.ver}{A 3-dimensional array for observations.}

\item{threshold}{The numeric threshold for computing the threat score.
Observation values larger than or equal to the threshold will be
converted to 1. Analog values will not be compared to this threshold value because
\strong{it is assumed that the ensemble function will convert analog ensemble to binary values}.}

\item{ensemble.func}{A function to convert the ensemble members (the 4th
dimension of analogs) into \strong{a binary}.}

\item{...}{Extra parameters for the ensemble.func.}
}
\description{
RAnEn::verifyThreatScore computes the threat score for a specific threshold.
The formula is adopted from
\href{https://www.wpc.ncep.noaa.gov/html/scorcomp.shtml}{WPC Verification page}.
}
