% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeConfig.R
\name{writeConfig}
\alias{writeConfig}
\title{RAnEn::writeConfig}
\usage{
writeConfig(config, file, overwrite = F)
}
\arguments{
\item{config}{The successfully configured Configuration object.}

\item{file}{The output file name}
}
\description{
RAnEn::writeConfig writes a Configuration class to multiple files, usually
observations and forecasts into NetCDF files and the rest configuration parameters
into a configuration file. The Configuration input should already be successfully
configured.
}
\author{
Weiming Hu \email{weiming@psu.edu}
}
