% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeConfiguration.R
\name{writeConfiguration}
\alias{writeConfiguration}
\title{RAnEn::writeConfiguration}
\usage{
writeConfiguration(
  config,
  obs.par.names,
  fcsts.par.names,
  xs,
  ys,
  file.prefix = "",
  folder = "./",
  silent = F
)
}
\arguments{
\item{config}{The successfully configured Configuration object.}

\item{obs.par.names}{The names of observation variables.}

\item{fcsts.par.names}{The names of forecast variables.}

\item{xs}{Xs of stations.}

\item{ys}{Ys of stations.}

\item{file.prefix}{File prefix for output files.}

\item{folder}{Output folder.}

\item{silent}{Whether to be silent.}
}
\description{
RAnEn::writeConfiguration writes a Configuration class to multiple files, usually
observations and forecasts into NetCDF files and the rest configuration parameters
into a configuration file. The Configuration input should already be successfully
configured.
}
\author{
Weiming Hu \email{weiming@psu.edu}
}
