% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeNcVar.R
\name{writeNcVar}
\alias{writeNcVar}
\title{RAnEn::writeNcVar}
\usage{
writeNcVar(variable_name, variable_data, dim_names, nc_file, verbose = F)
}
\arguments{
\item{variable_name}{The name of the variable}

\item{variable_data}{The data for the variable. It can be a vector, a matrix, or an array.}

\item{dim_names}{The dimension names associated with the variable.}

\item{nc_file}{The complete NetCDF file path}

\item{verbose}{Whether to be verbose}
}
\description{
RAnEn::writeNcVar can write a variable to a new NetCDF file if the file does not exist.
If the fist exists, RAnEn::writeNcVar tries to append the variable to the existing file.
}
\details{
This is a more general function than \code{RAnEn::writeNetCDF} because \code{RAnEn::writeNetCDF}
only takes a \code{Forecasts} or an \code{Observations}. But \code{RAnEn::writeNcVar} takes any numeric
arrays (including vectors, matrices, and arrays) and write them to an NetCDF file. This
can be helpful when you are saving \code{AnEn} results from \code{RAnEn::generateAnalogs}.
}
