% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeNetCDF.R
\name{writeNetCDF}
\alias{writeNetCDF}
\title{RAnEn::writeNetCDF}
\usage{
writeNetCDF(file.type, obj, file.out, global.attrs = NULL,
  nchars.max = 50)
}
\arguments{
\item{file.type}{The type of the object to be written to a file. It only supports
Observations for now.}

\item{nchars.max}{The maximum length of string.}

\item{global.attr}{A list with names and values for global attributes. The values
will be written as strings. Please make sure list members have names.}
}
\value{
A boolean for whether the writing is complete.
}
\description{
RAnEn::writeNetCDF writes a formatted object into a NetCDF file. The output NetCDF
file can be used by C++ utilities.
}
\details{
The formatted object for \code{Observations} is an R list object with the following members:
\itemize{
\item ParameterNames
\item StationNames (optional)
\item Xs
\item Ys
\item Times
\item Data
}
}
\author{
Weiming Hu \email{weiming@psu.edu}
}
