% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasCorrection.R
\name{biasCorrection}
\alias{biasCorrection}
\title{RAnEn::biasCorrection}
\usage{
biasCorrection(..., method = "insitu")
}
\arguments{
\item{...}{Paramters that are passed to the subsequent function.}

\item{method}{Which bias correction method to use. It can be one of
'insitu' and 'nnet'.}
}
\value{
An AnEn object.
}
\description{
RAnEn::biasCorrection
}
\author{
Weiming Hu \email{weiming@psu.edu}

This is a wrapper function for different bias correction methods.

Currently, it supports the following methods:
- Insitu bias correction \code{\link{RAnEn::biasCorrectionInsitu}};
- Neural network bias correction \code{\link{RAnEn::biasCorrectionNeuralNet}};
}
