% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasCorrection.R
\name{biasCorrection}
\alias{biasCorrection}
\title{RAnEn::biasCorrection}
\usage{
biasCorrection(AnEn, config, forecast.ID, group.func = mean, ...,
  keep.bias = F, show.progress = T)
}
\arguments{
\item{AnEn}{An AnEn object.}

\item{config}{The configuration object used to generate the AnEn results.}

\item{forecast.ID}{Which parameter in the forecasts should be used to generate bias.}

\item{group.func}{A function that summarizes all the member values
into one value. For example, it can be 'mean', 'median', 'max', or
'min'.}

\item{...}{Extra parameters to function group.func.}

\item{keep.bias}{A logical for whether to keep the bias array}

\item{show.progress}{A logical for whether to show the progress bar}
}
\value{
An array or a list depending on whether keep.bias is selected.
}
\description{
RAnEn::biasCorrection
}
\author{
Weiming Hu \email{weiming@psu.edu}

Statical bias correction function for analog ensembles.
}
