% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasCorrection.R
\name{biasCorrection}
\alias{biasCorrection}
\title{RAnEn::biasCorrection}
\usage{
biasCorrection(..., method = "insitu")
}
\arguments{
\item{...}{Paramters that are passed to the subsequent function.}

\item{method}{Which bias correction method to use. It can be one of
'insitu' and 'nnet'.}
}
\value{
An AnEn object.
}
\description{
This is a wrapper function for different bias correction methods.
}
\details{
Currently, it supports the following methods:

\itemize{
  \item Insitu bias correction \code{\link{RAnEn::biasCorrectionInsitu}};
  \item Neural network bias correction \code{\link{RAnEn::biasCorrectionNeuralNet}};
}
}
\examples{
\dontrun{
config <- generateConfiguration('independentSearch')

#
# Set up your configuration ...
#

AnEn <- generateAnalogs(config)
AnEn <- biasCorrection(AnEn, config, forecast.ID = config$observation_id, group.func = mean, na.rm = T,
                       keep.bias = T, show.progress = F, overwrite = F, method = 'insitu')
AnEn <- biasCorrection(AnEn, config, keep.bias = T, keep.model = T, group.func = mean,
                       na.rm = T,  show.progress = F, overwrite = F, parallel = T,
                       num.cores = 5, method = 'nnet')
}

}
\author{
Weiming Hu \email{weiming@psu.edu}
}
