% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_analogs.R
\name{compute_analogs}
\alias{compute_analogs}
\title{RAnEn::compute_analogs}
\usage{
compute_analogs(
  forecasts,
  observations,
  circulars = vector(mode = "numeric"),
  weights = NULL,
  stations_ID = NULL,
  test_ID_start,
  test_ID_end,
  train_ID_start = 1,
  train_ID_end = NULL,
  observation_ID = NULL,
  members_size,
  rolling = 0,
  quick = T,
  cores = 0,
  search_extension = F,
  xs = NULL,
  ys = NULL,
  distance = 0,
  num_nearest = 0,
  observation_from_extended_station = F,
  recompute_sd_for_extended_station = F,
  output_search_stations = F,
  output_metric = F,
  verbose = 1,
  parameter_ID = NULL,
  max_par_nan = 0,
  max_flt_nan = 0
)
}
\description{
Deprecated function. Please see \code{\link{generateAnalogs}}.
}
