% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateAnalogs.R
\name{generateAnalogs}
\alias{generateAnalogs}
\title{RAnEn::generateAnalogs}
\usage{
generateAnalogs(configuration)
}
\arguments{
\item{configuration}{A configuration object generated by the function
\code{\link{generateConfiguration}}. As a good practice, the configuration object
should be validated by the function \code{\link{validateConfiguration}} before
usage.}
}
\value{
An AnEn list with customized members. The list member \bold{analogs} 
is a 5-dimensional array that contains values and indices of ensemble
members from the variable observations. The dimensions are 
\bold{analogs[test stations, test days, FLT, members, ...]}
where FLT stands for forecast lead time. The size of the fifth dimension
is 3, storing the value [,,,,1], station index [,,,,2], and day index
[,,,,3]. The list member \bold{similarity} stores the similarity metrics.
The dimension indication is the same with the member analogs. The list
member \bold{searchStations} stores the search station index. The list member
\bold{mapping} stores the mapping mechanim from the time and the flt in
forecasts to the time in observations.
}
\description{
RAnEn::generateAnalogs returns the generated analog ensembles with either 
ensemble values and/or ensemble index. It can also return a number of 
auxiliary information, for example, similarity metrics and search stations.
These can be customized in the configuration object generated by the
function \code{\link{generateConfiguration}}.
}
\author{
Weiming Hu \email{weiming@psu.edu}
}
