% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateAnalogs.R
\name{generateAnalogs}
\alias{generateAnalogs}
\title{RAnEn::generateAnalogs}
\usage{
generateAnalogs(config)
}
\arguments{
\item{configuration}{A configuration object generated by the function
\code{\link{generateConfiguration}}.}
}
\value{
An AnEn list with customized members. Please check details in the
detail secion.
}
\description{
RAnEn::generateAnalogs returns an AnEn class object. It usually includes the
member \code{analogs} and other members as specified in the configuration.
}
\details{
The return value is an AnEn list.

The list member \strong{analogs} is a 5-dimensional array that contains values
and indices of ensemble members from the observations. The dimensions are
\strong{analogs[test stations, test days, FLT, members, ...]}
where FLT stands for forecast lead time. The size of the fifth dimension
is 3, storing the value [,,,,1], station index [,,,,2], and observation
time index [,,,,3]. Negative observation time indices are typically associated with
NA ensemble members and these negative values show the reason why a
particular member has an NA value. Please refer to
\href{https://weiming-hu.github.io/AnalogsEnsemble/CXX/class_an_en.html#a8e2b88cda5cc9fce8ea4703a3236719c}{AnEn::selectAnalogs}.

The list member \strong{similarity} stores the similarity metrics.
The dimension indication is the same with the member \strong{analogs}.
The size of the fifth dimension is 3, storing the similarity value
[,,,,1], station index [,,,,2], and forecast time index [,,,,3].
Negative forecast time indices are typically associated with NA
similarity values and these nagetive values show the reason why a particular
similarity value is NA. Please refer to
\href{https://weiming-hu.github.io/AnalogsEnsemble/CXX/class_an_en.html#a787e43e7bfdaa830bf70fa6710731614}{AnEn::computeSimilarity}.

The list member \strong{searchStations} stores the search station index.

The list member \strong{mapping} stores the lookup table mapping from
forecast times and flts to observation times. This table can alos be generated
using \code{\link{generateTimeMapping}}.
}
\author{
Weiming Hu \email{weiming@psu.edu}
}
