% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateConfiguration.R
\name{generateConfiguration}
\alias{generateConfiguration}
\title{RAnEn::generateConfiguration}
\usage{
generateConfiguration(mode)
}
\arguments{
\item{mode}{A character string of either 'independentSearch' or
'extendedSearch'.}
}
\value{
A configuration list with all the parameters you need to specify.
Please don't change the mode of the configuration mode. If you need a configuration
with a different mode, you need to call the function again with the desired mode.

Please find detailed terminology explanation in the
\href{https://weiming-hu.github.io/AnalogsEnsemble/2018/10/06/vocabulary.html}{Vocabulary post}.
If you still can't find helpful information or can't understand the parameters, please open a ticket.

For configuration with mode 'independentSearch':
\itemize{
\item mode: 'independentSearch' indicates only searching the forecasts at the current location.
\item test_forecasts: The forecasts for which anlog ensembles will be generated.
\item search_forecasts: The forecasts from which analogs are searched.
\item search_times: Time information for search forecasts.
\item search_flts: FLT information for search forecasts.
\item search_observations: The observations from which analog ensemble members are selected.
\item observation_times: Time information for search observations.
\item observation_id: The parameter index for which the observation parameter to use. For example,
if you have several observation parameters in search observations, this parameter tells the
package which observation parameter to use.
\item num_members: The number of ensemble member to keep in analog ensembles.
\item circulars: The index of forecast parameters in search forecasts that are circular.
\item weights: The weight for forecast parameters in search forecasts.
\item quick: Whether to use quick sort algorithm.
\item preserve_real_time: Whether to preserve the real time information in returned AnEn.
\item preserve_similarity: Whether to preserve the similarity matrix in returned AnEn.
\item preserve_mapping: Whether to preserve the mapping matrix in return AnEn.
\item verbose: Verbose level. You can find settings for different values in the
\href{https://weiming-hu.github.io/AnalogsEnsemble/CXX/class_an_en.html#a25984b953516a987e2e9eb23048e5d60}{C++ documentation verbose}.
\item time_match_mode: This is the method used to computed the mapping matrix between forecast times/flts and observation times.
0 for strict search (return error when matching observation times cannot be found) and 1 for loose search.
\item max_par_nan: The number of NAN values allowed when computing similarity across different parameters.
Set it to NA to allow any number of NAN values.
\item max_flt_nan: The number of NAN values allowed when computing FLT window averages.
Set it to NA to allow any number of NAN values.
}

For configuration with mode 'extendedSearch': (skipping the aforementioned parameters)
\itemize{
\item mode: 'extendedSearch' indicates search the forecasts from both the current location and the nearby locations.
\item test_stations_x: X coordinates of test stations/grid points.
\item test_stations_y: Y coordinates of test stations/grid points.
\item search_stations_x: X coordinates of search stations/grid points.
\item search_stations_y: Y coordinates of search stations/grid points.
\item max_num_search_stations: The maximum search stations to keep in the similarity matrix. The use of this parameter
will only affect the number of search stations to keep in the similarity matrix. This won't affect the correctness
of nearby station selection. This parameter is used in case of insufficient memeory. A good rule of thumb would be
setting it to num_nearest when num_nearest is used, or when distance is used, setting it to an estimated number that
is slightly bigger than the expexted number of stations that will be found within the distance. Again, even if you
set it to a number that is wierdly off, the selected nearby stations will be gauranteed to fulfill your requirement.
\item num_nearest: The number of nearest neighbors.
\item distance: The distance to search for nearby stations/grid points.
\item preserve_search_stations: Whether to preserve the search station list in returned AnEn.
}
}
\description{
RAnEn::generateConfiguration generates a templated configuration object to
be passed to the function \code{\link{generateAnalogs}} instead of passing various
parameters one by one. This object can be validated before use by the function
\code{\link{validateConfiguration}}.
}
\author{
Weiming Hu \email{weiming@psu.edu}
}
