% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateConfiguration.R
\name{generateConfiguration}
\alias{generateConfiguration}
\title{RAnEn::generateConfiguration}
\usage{
generateConfiguration(mode)
}
\arguments{
\item{mode}{A character string of either 'independentSearch' or
'extendedSearch'.}
}
\description{
RAnEn::generateConfiguration generates a templated configuration object to
be passed to the function \code{\link{generateAnalogs}} instead of passing various
parameters one by one. This object can be validated before use by the function
\code{\link{validateConfiguration}}.
}
\author{
Weiming Hu \email{weiming@psu.edu}
}
