% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAnalogTimeSeries.R
\name{plotAnalogTimeSeries}
\alias{plotAnalogTimeSeries}
\title{RAnEn::plotAnalogTimeSeries}
\usage{
plotAnalogTimeSeries(start.time, end.time, i.station, obs.id, obs.times,
  obs.data, anen.times, anen.flts, anen.data, fcst.id = NULL,
  fcst.times = NULL, fcst.flts = NULL, fcst.data = NULL,
  max.flt = 82800, par.name = "", return.data = F)
}
\arguments{
\item{start.time}{A POSIXct for the start time.}

\item{end.time}{A POSIXct for the end time.}

\item{i.station}{The station index to visualize. This can
be a single scalar so that the same station index for
observations, AnEn, and forecasts is used; or it can be
a named vector like this:
\code{i.station = c('obs' = 4, 'anen' = 15, 'fcst' = 10)} or
\code{i.station = c('obs' = 4, 'anen' = 15)}.}

\item{obs.id}{The observation parameter ID.}

\item{obs.times}{A vector for observation times. Usually
this is \code{observations$Times}.}

\item{obs.data}{An array for observation data. Usually
this is \code{observations$Data}.}

\item{anen.times}{A vector for AnEn times. Usually
this is \code{config$test_time_compares}.}

\item{anen.flts}{A vector for AnEn flts. Usually
this is \code{config$flts}.}

\item{anen.data}{An 4-dimensional array for analogs.}

\item{fcst.id}{The forecast parameter ID.}

\item{fcst.times}{A vector for forecast times. Usually
this is \code{forecasts$Times}.}

\item{fcst.flts}{A vector for forecast flts. Usually
this is \code{forecasts$flts}.}

\item{fcst.data}{An array for forecast data. Usually
this is \code{forecasts$Data}.}

\item{max.flt}{The cut off point value for both
forecasts and AnEn lead times.}

\item{par.name}{The variable name to be shown on figures.}

\item{return.data}{Whether to return the plot data
so that you can generate your own plots. This will
suppress plotting the figure inside this function.}
}
\value{
If \code{return.data} is TRUE, a list with two
data frames is returned; otherwise, it returns a ggplot
object.
}
\description{
RAnEn::plotAnalogTimeSeries plots forecasts and observations
for comparison over time.
}
\details{
Sometimes, it is helpful to plot forecasts and observations
for a long time series. However, this is always limited
by forecast lead times because the lead times are usually
for several days. What if we want to plot a time series that
spans longer than that?

This function generates a plot for that purpose. It simply
stacks forecasts from different days and all forecasts lead
times. Overlapping lead times will be cut off and earlier
lead times will be preferred and kept.
}
