% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.AnEn.R
\name{print.AnEn}
\alias{print.AnEn}
\title{RAnEn::print.AnEn}
\usage{
\method{print}{AnEn}(x)
}
\arguments{
\item{x}{an AnEn object}
}
\description{
RAnEn::print.AnEn
}
\examples{
AnEn <- list()
class(AnEn) <- 'AnEn'
print(AnEn)

}
\author{
Weiming Hu \email{weiming@psu.edu}

Overload print function for class AnEn
}
