% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Configuration.R
\name{print.Configuration}
\alias{print.Configuration}
\title{RAnEn::print.Configuration}
\usage{
\method{print}{Configuration}(x)
}
\arguments{
\item{x}{a Configuration object}
}
\description{
Overload print function for class Configuration
}
\examples{
config <- generateConfiguration('independentSearch')
print(config)

}
\author{
Weiming Hu \email{weiming@psu.edu}
}
