% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Templates.R
\name{print.Forecasts}
\alias{print.Forecasts}
\title{RAnEn::print.Forecasts}
\usage{
\method{print}{Forecasts}(x)
}
\arguments{
\item{x}{A Forecasts object}
}
\description{
RAnEn::print.Forecasts print a Forecasts object.
}
