% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAnEn.R
\docType{package}
\name{RAnEn-package}
\alias{RAnEn}
\alias{RAnEn-package}
\title{RAnEn::RAnEn}
\description{
The Analog Ensemble Package computes and manages analog ensembles. It is implemented in C++ and exported to R interface.
}
\details{
RAnEn is the R package implementing the Analog Ensemble method and more with data processing and visualization tasks.

Currently the available functions are:
- \code{\link{biasCorrection}}
- \code{\link{generateAnalogs}}
- \code{\link{generateCircleCoords}}
- \code{\link{generateConfiguration}}
- \code{\link{plotAnEnMap}}
- \code{\link{plotAnEnTimesSeries}}
- \code{\link{validateConfiguration}}
}
\author{
Weiming Hu \email{weiming@psu.edu}
}
\keyword{Analog}
\keyword{Bias}
\keyword{Correction}
\keyword{Ensemble,}
\keyword{Forecasts,}
\keyword{Probabilistic}
\keyword{Weather}
