% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAnEn.R
\docType{package}
\name{RAnEn-package}
\alias{RAnEn}
\alias{RAnEn-package}
\title{RAnEn::RAnEn}
\description{
The Analog Ensemble Package computes analog ensembles. It is implemented in C++ and exported to R interface.
}
\details{
RAnEn is the R package implementing the Analog Ensemble method and more with data processing and visualization tasks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://weiming-hu.github.io/AnalogsEnsemble/}
  \item \url{https://github.com/Weiming-Hu/AnalogsEnsemble}
  \item Report bugs at \url{https://github.com/Weiming-Hu/AnalogsEnsemble/issues}
}

}
\author{
Weiming Hu \email{weiming@psu.edu}
}
\keyword{Analog}
\keyword{Bias}
\keyword{Correction}
\keyword{Ensemble,}
\keyword{Forecasts,}
\keyword{Probabilistic}
\keyword{Weather}
