% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAnEn.R
\name{readAnEn}
\alias{readAnEn}
\title{RAnEn::readAnEn}
\usage{
readAnEn(file.analogs, file.similarity = NULL, verbose = F, split = 0)
}
\arguments{
\item{file.analogs}{The path to analog NetCDF file.}

\item{file.similarity}{The path to similarity NetCDF file.}

\item{verbose}{Whether to print progress.}

\item{split}{The number of times to split the longest dimension when reading.
For example, during the reading of analogs with 5 dimensions, if \code{split} is
3, the longest dimension of analogs will be divided 3 times into 4 chunks.
This is useful when your file cannot be read at once.}
}
\value{
An AnEn class object.
}
\description{
RAnEn::readAnEn can read AnEn results from NetCDF files that are generated
from analogGenerator. It supports reading analogs and similarity results.
}
\author{
Weiming Hu \email{weiming@psu.edu}
}
