% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readConfig.R
\name{readConfig}
\alias{readConfig}
\title{RAnEn::readConfig}
\usage{
readConfig(config.file)
}
\arguments{
\item{config.file}{The full path to the configuration file (.cfg)}
}
\value{
A config
}
\description{
RAnEn::readConfig reads a config file, usually generated by RAnEn::writeConfig
}
\author{
Weiming Hu \email{weiming@psu.edu}
}
