% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readConfig.R
\name{readConfig}
\alias{readConfig}
\title{RAnEn::readConfig}
\usage{
readConfig(config.file, parameter, return.numeric = F)
}
\arguments{
\item{config.file}{The full path to the configuration file (.cfg)}

\item{parameter}{The name of the configuration parameter to read.}

\item{return.numeric}{If this is TRUE, the returned vector will be converted to 
a numeric vector. If this is FALSE, the returned vector will be a character vector.}
}
\value{
A vector for the values of the specified configuration parameter.
}
\description{
RAnEn::readConfig reads a specific parameter from a configuration file (.cfg). The
configuration file is usually used by the C++ utility.
}
\author{
Weiming Hu \email{weiming@psu.edu}
}
