% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNc.R
\name{readNc}
\alias{readNc}
\title{RAnEn::readNc}
\usage{
readNc(file, origin = "1970-01-01", tz = "UTC", root_group_only = T)
}
\arguments{
\item{file}{The file to read}

\item{origin}{The origin of time to be used in as.POSIXct}

\item{tz}{The time zone of time to be used in as.POSIXct}

\item{root_group_only}{Whether to read only root group variables}
}
\description{
RAnEn::readNc reads an NetCDF file as a list
}
\author{
Weiming Hu \email{weiming@psu.edu}
}
