% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNc2.R
\name{readNc2}
\alias{readNc2}
\title{RAnEn::readNc}
\usage{
readNc2(file, recursive = F)
}
\arguments{
\item{file}{The file to read}

\item{recursive}{Whether to read variables recursively}
}
\description{
RAnEn::readNc reads an NetCDF file as a list with \code{RNetCDF} interface.
}
\author{
Weiming Hu \email{weiming@psu.edu}
}
