% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toMemberForecasts.R
\name{toMemberForecasts}
\alias{toMemberForecasts}
\title{RAnEn::toMemberForecasts}
\usage{
toMemberForecasts(
  forecasts,
  sims.time.index,
  sims.stations.index = NULL,
  stations.index = 1,
  test.times.index = 1,
  flts.index = 1,
  members.index = NULL,
  parameters.index = NULL,
  verbose = T
)
}
\description{
RAnEn::toMemberForecasts
}
\author{
Weiming Hu \email{weiming@psu.edu}

Laura Clemente-Harding \email{laura@psu.edu}
}
