% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val2Col.R
\name{val2Col}
\alias{val2Col}
\title{RAnEn::val2Col}
\usage{
val2Col(x, min = min(x, na.rm), max = max(x, na.rm), na.rm = F, col)
}
\arguments{
\item{x}{A vector of values.}

\item{min}{The minimum value.}

\item{max}{The maximum value.}

\item{na.rm}{Whether to remove NA values.}

\item{col}{A vector of colors to choose from.}
}
\description{
RAnEn::val2Col maps each value to a color in the input color sequence.
}
\author{
Guido Cervone \email{cervone@psu.edu}

Martina Calovi \email{mxc895@psu.edu}

Laura Clemente-Harding \email{laura@psu.edu}
}
