% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifySpreadSkill.R, R/verifySpreadSkillGuido.R
\name{verifySpreadSkill}
\alias{verifySpreadSkill}
\title{RAnEn::verifySpreadSkill}
\usage{
verifySpreadSkill(anen.ver, obs.ver, boot = F, R = 1000, na.rm = T,
  keep.cor = F)

verifySpreadSkill(anen.ver, obs.ver, boot = F, R = 1000, na.rm = T,
  keep.cor = F)
}
\arguments{
\item{anen.ver}{A 4-dimensional array. This array is usually created from the \code{value} column of
the \code{analogs} member in the results of \code{\link{generateAnalogs}}. The dimensions should be
\code{[stations, times, lead times, members]}.}

\item{obs.ver}{A 3-dimensional array. The dimensions should be \code{[stations, times, lead times]}.
You can generate the array using \code{\link{alignObservations}}.}

\item{boot}{Whether to use bootstrap.}

\item{R}{The number of bootstrap replicates. Used by the function \code{boot::boot}.}

\item{na.rm}{Whether to remove NA values.}

\item{keep.cor}{Whether to keep all spread and errors.}

\item{anen.ver}{A 4-dimensional array. This array is usually created from the \code{value} column of
the \code{analogs} member in the results of \code{\link{generateAnalogs}}. The dimensions should be
\code{[stations, times, lead times, members]}.}

\item{obs.ver}{A 3-dimensional array. The dimensions should be \code{[stations, times, lead times]}.
You can generate the array using \code{\link{alignObservations}}.}

\item{boot}{Whether to use bootstrap.}

\item{R}{The number of bootstrap replicates. Used by the function \code{boot::boot}.}

\item{na.rm}{Whether to remove NA values.}

\item{keep.cor}{Whether to keep all spread and errors. This might generate large data.}
}
\value{
A list with the verificatin results. The member \code{mean} is the average correlation
across all forecast cases; the member \code{flt} is the average correlation broken up by each
lead time; the member \code{flt.boot} contains bootstrap information; the member \code{cor} contains
all spread-error pairs that are used to calculate correlation.
}
\description{
RAnEn::verifySpreadSkill calculates the spread skill.

RAnEn::verifySpreadSkill calculates the spread skill correlation. The skill is calculated
as the squared differences between observations and forecasts. The spread is the variance
of a particular ensemble. Then correlation is calculated between the error and the
variance.
}
\author{
Weiming Hu \email{weiming@psu.edu}

Weiming Hu \email{weiming@psu.edu}
}
