% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifySpreadSkill.R
\name{verifySpreadSkill}
\alias{verifySpreadSkill}
\title{RAnEn::verifySpreadSkill}
\usage{
verifySpreadSkill(anen.ver, obs.ver, boot = F, R = 1000,
  intervals = 13, na.rm = T)
}
\description{
Underdevelopment.
}
